+++ app/assets/scripts/minireel/editor.js
@@ -2161,6 +2161,10 @@ VideoCardController           , c6embed) {
                             .on('error', $emitError);
                     }
 
+                    scope.disableTrimmer = function() {
+                        return (scope.service === 'vine');
+                    };
+
                     Object.defineProperties(scope, {
                         embedCode: {
                             get: function() {